pyproject.toml
[tool.poetry]
name = "enso-docs"
version = "0.1.0"
description = "Enso Standard Library documentation site generator."
authors = ["Enso Team <contact@enso.org>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.54.1"
black = "^20.8b1"
mypy = "^0.800"
PyExecJS = "^1.5.1"
pylint = "^2.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
